Scrimmage Commit
This commit is being made right after our first scrimmage. We were able
to get all the systems functional, though there is still a lot of work
to be done to improve them.
There are no changes in this commit except for small changes to get
things to work. One major concern of ours is trying to reduce the
amount of information being sent by the controllers so that the system
sare not overloaded by event handling. Other than that, we have to make
sure the PID loop of the arm is functional, which we had already done
last night but then it didn' t work today, make sure that we have the
right controller mapping, get rid of any lag the systems are having,
and create autonomous (which really won't be that tough once we have
the systems working well).
- Single shot removed, Autoshoot only.